Move network helpers into the resource
